requirements.txt
tensorflow-cpu==2.6.0
keras==2.6.0
numpy>=1.19.2,<1.24.0
pillow>=8.0.0
h5py>=3.1.0,<3.8.0
protobuf<4.0.0 